Decoding for small negative integers wasn't working properly.
The issue is that we were widening a byte value (unsigned) which would treat it as 255 when in reality it's a signed integer (-1).

Simplest fix is just to first convert it to the signed equivalent (int8) _then_ widen it to int64 or int32.

+++ sdk/messaging/azeventhubs/internal/go-amqp/internal/encoding/decode.go
@@ -893,7 +893,7 @@ func readLong(r *buffer.Buffer) (int64, error) {
 	switch type_ {
 	case TypeCodeSmalllong:
 		n, err := r.ReadByte()
-		return int64(n), err
+		return int64(int8(n)), err
 	case TypeCodeLong:
 		n, err := r.ReadUint64()
 		return int64(n), err
@@ -911,7 +911,7 @@ func readInt32(r *buffer.Buffer) (int32, error) {
 	switch type_ {
 	case TypeCodeSmallint:
 		n, err := r.ReadByte()
-		return int32(n), err
+		return int32(int8(n)), err
 	case TypeCodeInt:
 		n, err := r.ReadUint32()
 		return int32(n), err
